Fix url link. Move helpLink to correct place
